# node-api/docs/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": ""
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/create-card": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "org",
            "in": "query",
            "required": false,
            "type": "integer"
          },
          {
            "name": "channel",
            "in": "query",
            "required": false,
            "type": "integer"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 1
                },
                "cardNumber": {
                  "type": "string",
                  "example": "1234567890"
                },
                "bankAccountId": {
                  "type": "number",
                  "example": 1
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/remove-card": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "org",
            "in": "query",
            "required": false,
            "type": "integer"
          },
          {
            "name": "channel",
            "in": "query",
            "required": false,
            "type": "integer"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 1
                },
                "bankAccountId": {
                  "type": "number",
                  "example": 1
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/get-all-persons": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "org",
            "in": "query",
            "required": false,
            "type": "integer"
          },
          {
            "name": "channel",
            "in": "query",
            "required": false,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/get-person": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "org",
            "in": "query",
            "required": false,
            "type": "integer"
          },
          {
            "name": "channel",
            "in": "query",
            "required": false,
            "type": "integer"
          },
          {
            "name": "id",
            "in": "query",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/create-person": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "org",
            "in": "query",
            "required": false,
            "type": "integer"
          },
          {
            "name": "channel",
            "in": "query",
            "required": false,
            "type": "integer"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 1
                },
                "name": {
                  "type": "string",
                  "example": "John"
                },
                "surname": {
                  "type": "string",
                  "example": "Doe"
                },
                "email": {
                  "type": "string",
                  "example": "jdoe@gmail.com"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/search-persons-by-name": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "org",
            "in": "query",
            "required": false,
            "type": "integer"
          },
          {
            "name": "channel",
            "in": "query",
            "required": false,
            "type": "integer"
          },
          {
            "name": "name",
            "in": "query",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/search-persons-by-surname": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "org",
            "in": "query",
            "required": false,
            "type": "integer"
          },
          {
            "name": "channel",
            "in": "query",
            "required": false,
            "type": "integer"
          },
          {
            "name": "surname",
            "in": "query",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/search-persons-by-surname-email": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "org",
            "in": "query",
            "required": false,
            "type": "integer"
          },
          {
            "name": "channel",
            "in": "query",
            "required": false,
            "type": "integer"
          },
          {
            "name": "surname",
            "in": "query",
            "required": true,
            "type": "string"
          },
          {
            "name": "email",
            "in": "query",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/get-person-by-bank-account": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "org",
            "in": "query",
            "required": false,
            "type": "integer"
          },
          {
            "name": "channel",
            "in": "query",
            "required": false,
            "type": "integer"
          },
          {
            "name": "bankAccountId",
            "in": "query",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  }
}